requirements.txt
environs==9.5.0
python-binance==1.0.17
telethon==1.27.0
art==5.8